MAINT: Move golden_section_search into optimize file
